remove cdn folder to make things more clear
