--- src/cdn_parser/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CdnParserConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'cdn_parser'
