+++ src/eatery_blue_backend/urls.py
@@ -22,5 +22,4 @@ urlpatterns = [
     path("report/", include("report.urls")),
     path("alert/", include("alert.urls")),
     path("person/", include("person.urls")),
-    #path("wait_time/", include("wait_time.urls"))
 ]
